+++ decorators/__init__.py
@@ -1,0 +1,1 @@
+from .catch_usergrid_not_found import *
