+++ decorators/catch_usergrid_not_found.py
@@ -1,5 +1,5 @@
 import functools
-from usergrid.usergrid import UserGridException
+from exceptions import UserGridException
 
 
 def catch_usergrid_not_found_exception(return_value_on_exception=None):
