+++ exceptions/__init__.py
@@ -1,0 +1,1 @@
+from .usergrid_exception import UserGridException
